+++ translations/fr.py
@@ -402,9 +402,11 @@ def translateStartVmFR(window, vmname):
 
     window.pushButton.setText("Naviguer") # Browse
     window.pushButton_2.setText("Naviguer") # Browse
+    window.pushButton_6.setText("Naviguer") # Browse
     window.pushButton_5.setText("Set to system") # Set to system
     window.pushButton_3.setText("Start VM") # Start VM
     window.pushButton_4.setText("Annuler") # Cancel
+    window.checkBox.setText("Use RTC option") # Use RTC option
 
     # Combo box for boot
     i = 0