+++ translations/it.py
@@ -421,9 +421,11 @@ def translateStartVmIT(window, vmname):
 
     window.pushButton.setText("Sfoglia") # Browse
     window.pushButton_2.setText("Sfoglia") # Browse
+    window.pushButton_6.setText("Sfoglia") # Browse
     window.pushButton_5.setText("Imposta il sistema a") # Set to system
     window.pushButton_3.setText("Avvia VM") # Start VM
     window.pushButton_4.setText("Annulla") # Cancel
+    window.checkBox.setText("Use RTC option") # Use RTC option
 
     # Combo box for boot
     i = 0