+++ translations/ro.py
@@ -403,9 +403,11 @@ def translateStartVmRO(window, vmname):
 
     window.pushButton.setText("Browse") # Browse
     window.pushButton_2.setText("Browse") # Browse
+    window.pushButton_6.setText("Browse") # Browse
     window.pushButton_5.setText("Set to system") # Set to system
     window.pushButton_3.setText("Start VM") # Start VM
     window.pushButton_4.setText("Cancel") # Cancel
+    window.checkBox.setText("Use RTC option") # Use RTC option
 
     # Combo box for boot
     i = 0