+++ translations/uk.py
@@ -401,9 +401,11 @@ def translateStartVmUK(window, vmname):
 
     window.pushButton.setText("Огляд") # Browse
     window.pushButton_2.setText("Огляд") # Browse
+    window.pushButton_6.setText("Огляд") # Browse
     window.pushButton_5.setText("Set to system") # Set to system
     window.pushButton_3.setText("Запустіть VM") # Start VM
     window.pushButton_4.setText("Скасувати") # Cancel
+    window.checkBox.setText("Use RTC option") # Use RTC option
 
     # Combo box for boot
     i = 0