+++ CHANGELOG.md
@@ -7,6 +7,7 @@ You now need the latest version of your Linux distribution or at least Windows 1
 ## Changes compared to v1.2.0.5510
 
 - I forgot to actually assign the second ISO stuff. Now I realised and fixed it.
+- The translations are now more complete too.
 
 # EmuGUI v1.2.0.5510 "Garuka Pula" (based on v1.2.0.5509_rc2)